backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.8.2
python-jose[cryptography]==3.3.0
asyncio-mqtt==0.16.2
# Added dependencies:
# email-validator required by pydantic EmailStr fields.
email-validator==2.1.0.post1
# Pin paho-mqtt to 1.6.1 for asyncio-mqtt compatibility (avoids missing message_retry_set in paho 2.x).
paho-mqtt==1.6.1
